DLS-11425: [GMP] Apply Scala 3 Update to gmp-frontend

+++ app/repositories/GMPBulkSessionRepository.scala
@@ -46,7 +46,7 @@ class GMPBulkSessionRepository @Inject() (
         Indexes.ascending("lastModifiedIdx"),
         IndexOptions()
           .name("lastModifiedIdx")
-          .expireAfter(appConfig.cacheTtl, TimeUnit.SECONDS)
+          .expireAfter(appConfig.cacheTtl.toLong, TimeUnit.SECONDS)
       )
     )
   ){
